refactor: abstracted appError handler to middleware
