+++ apps/api/src/middleware/auth.ts
@@ -6,7 +6,6 @@ export const authMiddleware = bearerAuth({
     verifyToken: async (token, c) => {
         const usersService = container.resolve(UsersDbService);
         const user = await usersService.findOne({ apiKey: token });
-
         if (user) {
             c.set('currentUser', user);
             return true;
